Add support for multiple image queues (#8)
Being able to keep track of both "popular" and "user" queues simultaneously
should improve the perceived performance of this app from the user's
perspective. Previously, the current queue data was discarded each time the
user switched to a new tab, which meant waiting for new data from the API.

This change refactors the Vuex store to make certain properties (image array,
pending state) queue-dependent. Similarly, some of the mutations and actions
have been updated to become "queue-aware", typically defaulting to the current
active tab if a queue parameter is not provided.

Components have been updated to handle the new Vuex structure.

A few other changes that also made it in include:

* isAuthenticated and isAutoconfirmed have been pulled out of state and into
  getters, since they are technically returning derivative attributes
* attempted to consistently use $.extend with Vuex helpers; the format here is:
  $.extend( {}, mapWhatever( [] ), { localMethodOrProperty: function () {} } )
* Uses a watcher to ensure that fresh suggestions are copied into the ImageCard
  component's state the moment that its `image` prop changes
* Added method/computed property documentation
* Fleshed out the App.vue unit test

Change-Id: If58cc1e63fa62285a0a5a994903b98228aeb5df0

Co-authored-by: Eric Gardner <egardner@wikimedia.org>