Add local copy of i18n plugin to enable unit tests in CI (#10)
Until MediaWiki's i18n vue plugin is exposed as a node module, there is not a
great way to use this plugin when running unit tests in Jest. Previously we
have been reaching back up into the parent Mediawiki directory to find the file,
but this isn't possible when running JS unit tests in a stand-alone CI
environment. In order to allow for this, I've take the step to copy the plugin
code into our test folder for now. This can be removed once there is a better
way to consume this plugin outside of Mediawiki.

Change-Id: I0caf4dea7cc1fb879d887deb9e351e5e6280be28

Co-authored-by: Eric Gardner <egardner@wikimedia.org>