Handle app-specific toast logic, emit event when toast leaves
The base toast component will emit an event, including the component's
key, to the parent when the toast is removed. Optionally, we can
act on that event in the parent component, using the component key
if needed.

For our app, we need to make sure a new toast is used for each
notification so that we see the transition and so that the duration
is respected.

Change-Id: I5d74d4ea52c786cdcc133ce00e7a31f13dbf8b26