Refactor components to use Vuex instead of props for image & suggestion data
Since we're using Vuex here, we may as well rely on it wherever possible to
simplify questions about the "source of truth" in the application. There is
really no need to pass things like Image or Suggestions down from one component
to another as props any onger. This lets us eliminate some code inside the
components and solves our previous problem of how to deal with the scenario
where ImageCard receives suggestions as part of its props and then mutates a
local copy of  the data.

* Adds a new Vuex getter, currentImageSuggestions, for use in components
* Adds a new action and mutation for toggling the state of a single suggestion;
  ImageCard dispatches the toggleTagConfirmation action every time a tag is
  clicked, with the tag in question as a payload. The action looks up the index
  of this suggestion in the current image. This index is provided as an
  argument to the toggleSuggestion mutation which changes the state of the
  suggestion in state. User experience remains unchanged.
* the publishTags action no longer needs any payload at all, since all the
  necessary data is already in Vuex
* Reworks the tests to reflect the new data flow; a little more mocking is
  necessary for now.

Change-Id: I04960309c0f8abbf1c20dd44de7c3ed189d6fb47