Call existing OOUI onboarding dialog from Vue
Re-implementing the global process dialog and window manager functionality in
OOUI is currently beyond the scope of this project.

For now, it seems like the best approach will be to treat dialogs as a system
boundary; Vue components can launch dialogs by firing methods as normal, but
they will have no knowledge of or control over what happens in those dialogs
(unless an event is emitted that the component is listening for).

This change mostly just copies over some pre-existing widgets without changing
them (file import paths have been altered slightly but that's it).

Some modifications to linting and testing are also required to allow the two
frameworks to co-exist for now.

Change-Id: I6cfb23facc43c4a107bf849434faac8cf3cb5e27