Fix bug when toggling a suggestion when a non-displaying suggestion exists
It is possible that one or more suggestions for a given image cannot be
displayed to the user (because a label does not exist in their language, or a
wikidata item has moved/been deleted/etc.). When this happens we do not show
the suggestion to the user but it is still included in the original suggestions
array (and needs to be, for because of how we store reviewed state of labels in
the DB).

The problem is that we were mutating the suggestion "confirmed" state in Vuex
after locating the desired suggestion object using indexOf(). We were doing this
because Array.prototype.find() is not supported in IE. Unfortunatley the index
of a label in the displayed suggestions (the ones which we can show to the user)
may not always match the "real" index of the label in Vuex state if any
non-displayable suggestions are present.

In real-world terms, this leads to a confusing bug where the user clicks on
a label only to see the label *next to* whatever they clicked on get toggled.

This whole issue can be bypassed by simply using Array.prototype.find() to
locate the desired suggestion based on label text instead of array index.

In order to do this, a polyfill has been added for IE (kept in a separate
resource module called "ext.MachineVision.polyfill"). Tests have also been
updated.

Change-Id: I2835523666e0d3a1602fa2c6b93d957db295d7e9