Port MachineVision front-end to Vue.js
Over the past 2 months, Anne Tomasevich and I have been working to port over
the front-end of the MachineVision extension to Vue.js & Vuex. This commit
would replace the current, OOUI-based front-end (the contents of the
`/resources` folder, primarily) while leaving the rest of the code mostly
untouched.

A complete description of all changes is beyond the scope of this commit
message (see the fork at
https://github.com/egardner/mediawiki-extensions-MachineVision for a complete
look at everything that has been done here). Here is a summary:

* Adds some new resource modules to `ext.MachineVision` as dependencies: `vue`
  and `vuex` (both have been added to MW core).
* Re-organizes the `resources` directory – it is now structured as follows:
    resources/
    ├── base
    ├── config
    ├── components
    ├── icons
    ├── init.css
    ├── init.js
    ├── models
    ├── plugins
    ├── polyfills
    ├── store
    ├── style-variables.less
    └── widgets
* Most UI elements are now defined as single-file components in the components/ directory
* Vuex code lives in store (mutations, actions, state, etc)
* plugins are used to provide access to properties or methods across all
  components in the application; this extension includes two (a wrapper for
  mw.Api and an event logging method)
* Some widgets remain from the previous implementation because re-writing
  OOUI's ProcessDialog and WindowManager widgets was out of scope for this
  project; these are the only OOUI elements still in use here
* A polyfill for Array.prototype.find has been added in the polyfills/ folder.

In addition to the changes in the resources folder, new test scaffolding has
been added for front-end unit tests. We are using Jest for this because it is
able to handle Vue SFCs without requiring a Webpack build step. The
vue-test-utils library is used extensively here.

Co-authored-by: annet <atomasevich@wikimedia.org>

Bug: T248405
Change-Id: I90c1b95ff26c4b7dfb6fdb89157162cefe779b91